change amount migration data type to double

+++ database/migrations/2023_08_14_183648_create_employee_cashes_table.php
@@ -23,7 +23,7 @@ return new class extends Migration
 
             $table->foreignUuid('employee_id')->nullable();
 
-            $table->integer('amount')->default(0)->nullable();
+            $table->double('amount')->default(0)->nullable();
 
             $table->string('payer_name')->nullable();
